Implement terrain editor stubs, add TerrainSystem ECS, debloat/split files
- Implement UpdateTerrainMesh(): builds CPU-side vertex positions, normals,
  and triangle indices from heightmap data for rendering preview
- Implement UpdateTerrainCollision(): builds world-space height array for
  Bullet physics heightfield collision shapes
- Wire mesh/collision rebuild into terrain creation, loading, and generation
  tools (called on heightmap changes via dirty flags in Update())
- Add TerrainSystem ECS: processes TerrainComponent entities with LOD
  selection based on camera distance and heightmap bounds clamping
- Split TerrainEditor.h (696→145 lines) by extracting data structs to
  TerrainData.h (177 lines)
- Split TerrainEditor.cpp (1303→385 lines) into:
  - TerrainData.cpp (228 lines): brush, heightmap, splatmap implementations
  - TerrainEditorUI.cpp (355 lines): ImGui rendering methods
  - TerrainEditorTools.cpp (397 lines): tool application, noise, erosion
- All files now within 400-line (.cpp) and 200-line (.h) limits
