+++ SparkEngine/Source/Engine/ECS/Systems/TerrainSystem.h
@@ -1,0 +1,42 @@
+/**
+ * @file TerrainSystem.h
+ * @brief ECS system that processes TerrainComponent entities each frame
+ */
+
+#pragma once
+
+#include "ECSystems.h"
+#include "../Components/TerrainComponents.h"
+
+namespace Spark::ECS
+{
+
+    /**
+     * @class TerrainSystem
+     * @brief Updates terrain LOD and marks dirty terrains for mesh/collision rebuild.
+     *
+     * Each frame, iterates entities with TerrainComponent + Transform:
+     * 1. Selects LOD level based on camera-to-terrain distance.
+     * 2. Clamps heightmap data to configured min/max height bounds.
+     *
+     * ### Execution order
+     * Should run AFTER PhysicsUpdateSystem (so terrain collision reads updated positions)
+     * and BEFORE RenderSystem (so LOD selection is ready for draw submission).
+     */
+    class TerrainSystem : public ISystem
+    {
+      public:
+        void Update(World& world, float deltaTime) override;
+        const char* GetName() const override { return "TerrainSystem"; }
+
+        int GetActiveTerrainCount() const { return m_activeTerrainCount; }
+
+        /// Set the camera position used for LOD selection (updated by the player/camera system).
+        void SetCameraPosition(const XMFLOAT3& pos) { m_cameraPosition = pos; }
+
+      private:
+        int m_activeTerrainCount = 0;
+        XMFLOAT3 m_cameraPosition = {0, 0, 0};
+    };
+
+} // namespace Spark::ECS
